// cypress/enums/workType.ts
export enum WorkType {
    EMPLOYMENT = "EMPLOYMENT",
    ENTREPRENEUR = "ENTREPRENEUR",
    UNEMPLOYED = "UNEMPLOYED",
}

export enum OccupationType {
    SERVICES = "SERVICES",
    MANUAL = "MANUAL",
}

export enum UnemploymentType {
    MATERNITY_LEAVE = "Mateřská dovolená",
    STUDENT = "Student (nepracující)",
    UNEMPLOYED = "Nezaměstnaný",
    WOMAN_OR_MAN_IN_THE_HOUSEHOLD = "Žena (muž) v domácnosti",
}

export enum Occupation {
    OFFICE_WORKER = "Administrativní pracovník",
    LAWYER = "Advokát",
}