// cypress/enums/sportType.ts
export enum SportType {
   BADMINTON = "Badminton",
   TENNIS = "Tenis",
   FOOTBALL = "Fotbal",
   VOLLEYBALL = "Volejbal",
   SWIMMING = "Plavání",
}

export enum SportLevel {
    RECREATIONAL = "Rekreační",
    MEMBER_OF_CLUB = "Sportovní registrace – do krajské úrovně včetně nebo trenér",
    MEMBER_OF_CLUB_NATIONAL = "Sportovní registrace - celostátní a mezinárodní",
    PROFESSIONAL = "Profesionální",
}